requirements.txt
streamlit>=1.28.0
textblob>=0.17.1
pandas>=2.0.0
nltk>=3.8.1
groq>=0.4.0
python-dotenv>=1.0.0
